requirements.txt
flask
Gunicorn
pandas
plotly.express
os
render_template
request
jsonify